Fix callback calling syntax in sub arguments
In current Rakudo (2015.05) the current code didn't compile anymore, giving
the warning that 'Shape declaration with () is reserved;' and giving the
advice that one should either separate the subarguments with a colon (which
jnthn++ advised to do, however unfortunately didn't allow the code to
compile) or to add a space between the sub-sub-name and its argument.  Also,
having `Code &callback` was a bit doubled-up since the `&callback` was
returning a `Callable` anyway and as it turned out removing the `Code`
modifier allowed the code to compile and the test suite to run all tests bar
one.