Fix calls to SDL and SDL_mixer libraries
It seems that in 6.c there is no longer the need to prefix a library
name with 'lib'; this is prepended automatically.  In other words,
`native('libSDL')` should now be written as `native('SDL')`.  The tests
were complaining about not finding the `liblibSDL` and
`liblibSDL_mixer` libraries, hence this patch.

The compiler also warned not to return `Int` in some cases and that
`int32` should be used instead.  In the places where the compiler was
complaining, the return type has been updated appropriately.

These changes get the core tests working again and the mixer channel
tests at least compiling, however not running completely.