Rewrite all bindings, and expose libSDL's function/struct names
With this one will get all structures and functions that the C library offers
by doing `use SDL`. The sugar-coated classes will be in the SDLx namespace
only. But I've removed these for now, since the existing code was created
before NativeCall was able to unpack C Structures properly. So it was awkward
anyway and not much of a loss.